% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProjectHeader.R
\name{createProjectHeader}
\alias{createProjectHeader}
\title{Create a header for QC report}
\usage{
createProjectHeader(QCreportObject)
}
\arguments{
\item{QCreportObject}{QCreportObject}
}
\description{
Create a header for QC report
}
