% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQCreport.R
\name{createQCreport}
\alias{createQCreport}
\title{Create a pdf output from data slots on QCreportObject}
\usage{
createQCreport(QCreportObject, outputfile = NULL)
}
\arguments{
\item{QCreportObject}{Qcreport object}

\item{outputfile}{Name of the output pdf file}
}
\description{
Create a pdf output from data slots on QCreportObject
}
