% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQCreportObject.R
\name{createQCreportObject}
\alias{createQCreportObject}
\title{Create initial qcrepotobject structure from input variable}
\usage{
createQCreportObject(FILENAME_RDATA, projectdir, author = "", metaData_file,
  raw_path = NULL, InternalProjectRef = "", Dataset = "",
  Organisation = "", QC_label = "QC", Blank_label = "Blank",
  classLabel = "Class", excludeQC = c(1:5), assay = NULL)
}
\arguments{
\item{FILENAME_RDATA}{Rdata file containing XCMS output, object should be called xset}

\item{projectdir}{Folder where Rdata file is located and where output should be directed}

\item{author}{Character sring woth reports author name, optional}

\item{metaData_file}{Files in xlsx format containing sample names and class labels}

\item{raw_path}{File path where mzML files are located}

\item{InternalProjectRef}{Character string containing project reference, optional}

\item{Dataset}{Character string to describe data set, optional}

\item{Organisation}{Character string containing organisation name, optional}

\item{QC_label}{Label used for QC samples}

\item{Blank_label}{Label usef for QC samples}

\item{classLabel}{Columns name in meta data file to use for class labels, default value is Class}

\item{excludeQC}{Which QC samples to exclude from RSD\% calculations, signal batch correction and PCA plots}

\item{assay}{Names for analytical assay, if set to NULL name of the input Rdata file will be used}
}
\description{
Create initial qcrepotobject structure from input variable
}
