% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createXlsx.R
\name{createXlsx}
\alias{createXlsx}
\title{Create xlsx output file from XCMS output}
\usage{
createXlsx(QCreportObject)
}
\arguments{
\item{QCreportObject}{Qcreport object}
}
\description{
Create xlsx output file from XCMS output
}
