% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_variabilty_list.R
\name{do_variability_list}
\alias{do_variability_list}
\title{Calculate RSD\% or MAD values per sample group of input peakmatrix}
\usage{
do_variability_list(peak_data, classes, method = "RSD")
}
\arguments{
\item{peak_data}{peak matrix}

\item{classes}{Vector of class labels}

\item{method}{What method to use to summarise values. Supported methods are 
'RSD', 'MAD', 'median' and 'none'.}
}
\value{
List of RSD\% or MAD values for each feature for each sample group
}
\description{
Calculate RSD\% or MAD values per sample group of input peakmatrix
}
