% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_variability_plot.R
\name{do_variability_plot}
\alias{do_variability_plot}
\title{Plot violin plots from the dolist_object fucntion output}
\usage{
do_variability_plot(list_object, plotTitle = NULL, base_size = 12,
  subtitle = NULL, ylim = NULL)
}
\arguments{
\item{list_object}{OUtput of do_variability_list function}

\item{plotTitle}{Main tilte for the plot}

\item{base_size}{Font size for plot fonts}

\item{subtitle}{Subtitle to include in PCA plot}

\item{ylim}{If not NULL this value will be used for maximum of violin plots}
}
\value{
Ggplot object with plot(s)
}
\description{
Plot violin plots from the dolist_object fucntion output
}
