% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_variability_table.R
\name{do_variability_table}
\alias{do_variability_table}
\title{Create summary table of list object containing variabilty measures, for example output object of do_variability list function}
\usage{
do_variability_table(list_object, QC_label = "QC", Blank_label = "Blank")
}
\arguments{
\item{list_object}{OUtput of dolist_object function}

\item{QC_label}{Label used for QC samples. If set to NULL, assumes that no QC samples are present in data set}

\item{Blank_label}{Label used for Blank samples}
}
\value{
Table of list_object% values per group
}
\description{
Create summary table of list object containing variabilty measures, for example output object of do_variability list function
}
