% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doPCA.R
\name{doPCA}
\alias{doPCA}
\title{Perform PCA analysis ot single data set or list of data sets}
\usage{
doPCA(Data, classes, plotTitle = "PCA", PQN, mv_impute, glogScaling,
  scale = F, labels = "QC", qc_label, qc_shape = 17,
  base_size = 12, pccomp = c(1, 2), subtitle = NULL, loadings = F,
  loadingsCol = NULL)
}
\arguments{
\item{Data}{Data frame.}

\item{classes}{Vector of class labels.}

\item{plotTitle}{Character value to display as the plot title}

\item{PQN}{Can be set to T or F, to perform PQN normalisation}

\item{mv_impute}{T or F, indicates if missing value imputation has to be carried}

\item{glogScaling}{T or F, applie glog transformation to the given data}

\item{scale}{Perform UV scaling on data}

\item{labels}{Can be set to "QC" to label only QC samples. "none" to no include labels. If set to any other value will use column names of Data.}

\item{qc_label}{Label used for QC samples. If set to NULL, assumes that no QC samples are present in data set}

\item{qc_shape}{Shape symbol to use for QC samples}

\item{base_size}{Font size for plot fonts}

\item{pccomp}{PCA components to plot}

\item{subtitle}{Subtitle to include in PCA plot}

\item{loadings}{T or F, to include PCA loadings plot or not}

\item{loadingsCol}{Colors to use for loadings plot}
}
\value{
Ggplot object with plot(s)
}
\description{
Perform PCA analysis ot single data set or list of data sets
}
