% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQCreportObject.R
\name{file_type}
\alias{file_type}
\title{Check if input is a text file or xlsx file}
\usage{
file_type(path)
}
\arguments{
\item{path}{path to the file}
}
\description{
Check if input is a text file or xlsx file
}
