% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createXlsx.R
\name{formatIonIdentifiers}
\alias{formatIonIdentifiers}
\title{This function format ions identifiers}
\usage{
formatIonIdentifiers(variableMetadata, numDigitsRT = 0,
  numDigitsMZ = 0)
}
\arguments{
\item{variableMetadata}{vairable meta data object}

\item{numDigitsRT}{decimal digits for RT}

\item{numDigitsMZ}{number of deciaml digits for mz}
}
\description{
This function format ions identifiers
}
\author{
G. Le Corguille
}
