% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSummary.R
\name{get_file_extension}
\alias{get_file_extension}
\title{Return file extension from character vector if file paths}
\usage{
get_file_extension(file_paths)
}
\arguments{
\item{file_paths}{character(), character vector containing file paths}
}
\value{
character(), character vector with unique file extensions
}
\description{
Return file extension from character vector if file paths
}
