% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createXlsx.R
\name{getPeaklistW4M}
\alias{getPeaklistW4M}
\title{This function prepares peaklist object from xcms output}
\usage{
getPeaklistW4M(xset, intval = "into")
}
\arguments{
\item{xset}{XCMS object to process}

\item{intval}{Which value to use for peak area, see XCMS manual for more
details}
}
\description{
This function prepares peaklist object from xcms output
}
\author{
G. Le Corguille
}
