% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSummary.R
\name{locate_raw_files}
\alias{locate_raw_files}
\title{Locate raw LC/MS raw files on local files system and align with xcms output}
\usage{
locate_raw_files(QCreportObject)
}
\arguments{
\item{QCreportObject}{Qcreport object}
}
\description{
Locate raw LC/MS raw files on local files system and align with xcms output
}
