% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSummary.R
\name{meta_data_from_filenames}
\alias{meta_data_from_filenames}
\title{Generate meta data from sample filenames}
\usage{
meta_data_from_filenames(QCreportObject)
}
\arguments{
\item{QCreportObject}{Qcreport object}
}
\description{
Generate meta data from sample filenames
}
