% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValues.R
\name{missingValues}
\alias{missingValues}
\title{Create QCreport data object to asses missing values}
\usage{
missingValues(QCreportObject)
}
\arguments{
\item{QCreportObject}{Qcreport object}
}
\description{
Create QCreport data object to asses missing values
}
