% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzML.startTimeStamp.R
\name{mzML.startTimeStamp}
\alias{mzML.startTimeStamp}
\title{Extract measurement time stamp fro mzML data file}
\usage{
mzML.startTimeStamp(filename)
}
\arguments{
\item{filename}{Name of the mzML file}
}
\description{
Extract measurement time stamp fro mzML data file
}
