% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{Perform severl PCA models and create plots}
\usage{
PCA(QCreportObject)
}
\arguments{
\item{QCreportObject}{Qcreport object}
}
\description{
Perform severl PCA models and create plots
}
