% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{prepare_boxplot_multiple_batches}
\alias{prepare_boxplot_multiple_batches}
\title{Prepare data for ggplot barplot}
\usage{
prepare_boxplot_multiple_batches(vals, batch, class)
}
\arguments{
\item{batch}{Vector with sample batch labels}

\item{val}{Data frame to be processed}
}
\description{
Prepare data for ggplot barplot
}
