% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSDstatistics.R
\name{RSDstatistics}
\alias{RSDstatistics}
\title{Create QCreport data object to asses RSD}
\usage{
RSDstatistics(QCreportObject)
}
\arguments{
\item{QCreportObject}{Qcreport object}
}
\description{
Create QCreport data object to asses RSD
}
