% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createXlsx.R
\name{RTSecondToMinute}
\alias{RTSecondToMinute}
\title{This function convert if it is required the Retention Time in minutes}
\usage{
RTSecondToMinute(variableMetadata, convertRTMinute)
}
\arguments{
\item{variableMetadata}{vairable meta data object}

\item{convertRTMinute}{convert RT in seconds to minutes}
}
\description{
This function convert if it is required the Retention Time in minutes
}
\author{
G. Le Corguille
}
