% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSummaryPlots.R
\name{sampleSummaryPlots}
\alias{sampleSummaryPlots}
\title{Create plots of TIC chromatograms and general sample summary.}
\usage{
sampleSummaryPlots(QCreportObject)
}
\arguments{
\item{QCreportObject}{Qcreport object}
}
\description{
Create plots of TIC chromatograms and general sample summary.
}
