% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalBatchCorrection.R
\name{SignalBatchCorrection}
\alias{SignalBatchCorrection}
\title{Apply signal batch correction}
\usage{
SignalBatchCorrection(QCreportObject)
}
\arguments{
\item{QCreportObject}{Qcreport object}
}
\description{
Apply signal batch correction
}
