% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme_pub.R
\name{theme_Publication}
\alias{theme_Publication}
\title{Ggplot publication theme}
\usage{
theme_Publication(base_size = 14)
}
\arguments{
\item{base_size}{Ggplot font size}
}
\description{
This ggplot2 theme is modified from original by  Koundinya Desiraju.
Availaible from here: https://rpubs.com/Koundy/71792
}
