% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EIC_plots.R
\name{XCMS_EIC_plot}
\alias{XCMS_EIC_plot}
\title{Exctract and return ggplot2 object of EICs for specified xcms features}
\usage{
XCMS_EIC_plot(indexes, rawfiles, class, xset, Blank_label = "Blank",
  QC_label = "QC")
}
\arguments{
\item{indexes}{vector of feautre indices to extract}

\item{rawfiles}{List of MsnBase packages raw file objects}

\item{class}{Sample class labels}

\item{xset}{XCMS object from withc to extrat chromatograms}

\item{Blank_label}{Label used for Blank samples}

\item{QC_label}{Label used for QC samples}
}
\description{
Exctract and return ggplot2 object of EICs for specified xcms features
}
