% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSummary.R
\name{sampleSummary}
\alias{sampleSummary}
\title{Exctract and return ggplot2 object of EICs for specified xmcm features}
\usage{
sampleSummary(QCreportObject)
}
\arguments{
\item{QCreportObject}{Qcreport object
xcms::rtime(QCreportObject$xset, adjusted=FALSE, 
   bySample=TRUE)[msms_sample_hits] <- NULL
QCreportObject$xset@rt$corrected[msms_sample_hits] <- NULL}
}
\description{
Exctract and return ggplot2 object of EICs for specified xmcm features
}
